Refactor to make text support optional
