+++ avenger-wgpu/src/marks/mod.rs
@@ -1,9 +1,9 @@
+pub mod gradient;
+pub mod image;
 pub mod instanced_mark;
 pub mod multi;
 pub mod symbol;
-
-pub mod gradient;
-pub mod image;
+pub mod text;
 
 #[cfg(feature = "cosmic-text")]
-pub mod text;
+pub mod cosmic;
